Add files via upload
Data Description: Course Reviews

This dataset contains feedback data for online courses. Each entry represents a review provided by a student for a specific course lecture.

course_name: The name of the online course.
lecture_name: The title of the lecture within the course.
review_rating: The rating given by the student for the lecture, on a scale from 1 to 5 (1 being the lowest and 5 being the highest).
review_comment: Additional comments or feedback provided by the student about the lecture.